refactor(users): rename files and update folders structure of users repositories
