+++ src/users/repositories/implementations/index.ts
@@ -1,4 +1,1 @@
-export { TyeOrmOrganizersRepository } from './organizers.repository';
-export { TyeOrmParticipantsRepository } from './participants.repository';
-export { TyeOrmSpeakersRepository } from './speakers.repository';
-export { TyeOrmUsersRepository } from './users.repository';
+export * from './type-orm';
