+++ src/users/repositories/implementations/type-orm/index.ts
@@ -1,0 +1,4 @@
+export { TyeOrmOrganizersRepository } from './type-orm-organizers.repository';
+export { TyeOrmParticipantsRepository } from './type-orm-participants.repository';
+export { TyeOrmSpeakersRepository } from './type-orm-speakers.repository';
+export { TyeOrmUsersRepository } from './type-orm-users.repository';
