+++ src/users/repositories/implementations/type-orm/type-orm-participants.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
-import { Participant } from '../../entities';
+import { Participant } from '../../../entities';
 
 @Injectable()
 export class TyeOrmParticipantsRepository {
