+++ src/users/repositories/implementations/type-orm/type-orm-users.repository.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { User } from '../..//entities';
+import { User } from '../../../entities';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
