+++ src/users/repositories/index.ts
@@ -1,2 +1,2 @@
 export * from './interfaces';
-export * from './implementations';
+export * from './implementations/type-orm';
